Move main_window.py into RimSort.views
